fix chromedriver, and setup guimgr for open and close browser
