+++ Sdet.Auto.PageObjects/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Selenium.Support" version="3.0.1" targetFramework="net452" />
+  <package id="Selenium.WebDriver" version="3.0.1" targetFramework="net452" />
+</packages>
